Added Odds of Winning, broken images
Odds of winning is now displayed if there are entries to the giveaway. This adds a bit more transparency for the giveaway. If there are no entires, the
field is hidden and reformatted. Some broken images have been added and fixed. There is a new schema to assist with both the entries and the all giveaway buttons.